// frontend/src/api/generated/agents/agents.ts
export type getAgentApiV1AgentsAgentIdGetResponse200 = {
  data: AgentRead;
  status: 200;
};

export type getAgentApiV1AgentsAgentIdGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type getAgentApiV1AgentsAgentIdGetResponseSuccess =
  getAgentApiV1AgentsAgentIdGetResponse200 & {
    headers: Headers;
  };
export type getAgentApiV1AgentsAgentIdGetResponseError =
  getAgentApiV1AgentsAgentIdGetResponse422 & {
    headers: Headers;
  };

export type getAgentApiV1AgentsAgentIdGetResponse =
  | getAgentApiV1AgentsAgentIdGetResponseSuccess
  | getAgentApiV1AgentsAgentIdGetResponseError;